fix:fix reportClient not working bug when using push-gateway strategy.
